Clean up the cutlass benchmarking
